fixed subsystem and moved files
